Remove "sensors" from crate, converting them back to examples
Including the sensor support in this crate did not make sense, but
the code that exists is still pretty handy as a reference.  This commit
gets rid of all sensor traits and sensors from this crate and moves
the code to live in examples as a reference (pretty directly
pasted back into example files as a start).
