+++ src/lib.rs
@@ -27,10 +27,7 @@ extern crate nix;
 mod ffi;
 
 pub mod core;
-pub mod sensors;
 
 #[cfg(any(target_os = "linux", target_os = "android"))]
 pub mod linux;
-
-#[cfg(test)]
 pub mod mock;
